+++ js/exporter.ts
@@ -6,6 +6,7 @@ import { WidgetModel, WidgetView } from '@jupyter-widgets/base';
 import { unpack_models as deserialize } from '@jupyter-widgets/base';
 
 import { ELKViewerModel } from './display_widget';
+import { isEnabled } from './exporter_util';
 import { ELK_DEBUG, NAME, VERSION } from './tokens';
 
 import elkRawCSS from '!!raw-loader!../style/diagram.css';
@@ -27,7 +28,7 @@ const XML_HEADER = '<?xml version="1.0" standalone="no"?>';
 
 export class ELKExporterModel extends WidgetModel {
   static model_name = 'ELKExporterModel';
-  private _update_timeout: number;
+  private _update_timeout: ReturnType<typeof setTimeout> | null;
 
   static serializers = {
     ...WidgetModel.serializers,
@@ -57,7 +58,7 @@ export class ELKExporterModel extends WidgetModel {
   }
 
   get enabled(): boolean {
-    return this.get('enabled') || true;
+    return isEnabled(this.get('enabled'));
   }
 
   get viewer(): ELKViewerModel {