+++ tests/tools/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) 2024 ipyelk contributors.
+# Distributed under the terms of the Modified BSD License.